New interfacing refactor, fixing mass APIResponse interface issue exports all interfaces and updates embeds to match
