--- src/interfaces/_apiSlashCommandReq.ts
@@ -1,4 +1,0 @@
-export interface SlashCommandRequestInt {
-	command: string
-	payload: {};
-}
