+++ src/interfaces/_post.ts
@@ -1,0 +1,5 @@
+export interface PostInt {
+	id: string;
+	target: string;
+	body: string;
+}
