+++ src/interfaces/api/_CronfigResInt.ts
@@ -1,0 +1,6 @@
+export interface CronfigResInt {
+	message: string;
+	payload: {
+		cronfig: number[];
+	};
+}
