+++ src/interfaces/api/_CurrentCronfigResInt.ts
@@ -1,0 +1,7 @@
+export interface CurrentCronfigResInt {
+	message: string;
+	payload: {
+		cronfig: number[];
+		run_on_start: boolean;
+	};
+}
