+++ src/interfaces/api/_QueRoutineResInt.ts
@@ -1,0 +1,8 @@
+import type { PostInt } from '../../interfaces';
+
+export interface QueRoutineResInt {
+	message: string;
+	payload: {
+		posts: PostInt[];
+	};
+}
