+++ src/messages/embeds/_addQue.ts
@@ -1,16 +1,14 @@
 import { CommandInteraction, MessageEmbed } from 'discord.js';
 
-interface AddQueInfoInt {
-	id: string;
-	body: string;
+interface AddQueEmbedInt {
 	target: string;
+	body: string;
 }
-
-export const addQueEmbed = (interaction: CommandInteraction, { id, target, body }: AddQueInfoInt) => {
+export const addQueEmbed = (interaction: CommandInteraction, { target, body }: AddQueEmbedInt) => {
 	return new MessageEmbed()
 		.addFields(
 			{ name: 'Channel Target:', value: `${target}` },
-			{ name: 'Post ID:', value: `${id}` },
+			{ name: 'Post ID:', value: `${interaction.id}` },
 			{ name: 'Post Body:', value: `${body}` }
 		)
 		.setTitle('added a post to the queue:')
